Add Norse Bold font and adjust spacing and sizes
Went and measured the target image that I am suposed to be recreating,
and have started adusting CSS based on that. The logo has been moved
down slightly, the shaded bar made a little smaller, and the color has
been made darker. I added the Norse Bold font and changed the size, but
it's not displaying correctly.

On the right side, I adjusted the left padding to match, and put in the
appropriate sizes for the input boxes and the submit button. I also
changed the background colors slightly to match the image.