Add required/valid tags and fix H1 font
I added tags on the required fields (password and confirm password) so
that while the fields are invalid, a red "required" appears. When an
acceptible entry is detected, the red "required" changes to a green
"valid". We haven't learned how to use JS to validate that the confirm
password field matches the first password field, but if/when we get
there I would have the password field change to valid, and the confirm
field would have appropriate contextual messages.

I did also figure out that the custom font CSS was not pointing to the
correct folder. Fixed that, and the font displays.